+++ govtool/frontend/src/components/organisms/AutomatedVotingOptions.tsx
@@ -9,9 +9,9 @@ import {
 import { Typography } from "@atoms";
 import { ICONS } from "@consts";
 import { PendingTransaction } from "@context";
-import { useGetNetworkMetrics, useTranslation } from "@hooks";
+import { useGetNetworkTotalStake, useTranslation } from "@hooks";
 import { AutomatedVotingCard } from "@molecules";
-import { correctVoteAdaFormat, openInNewTab } from "@/utils";
+import { correctDRepDirectoryFormat, openInNewTab } from "@/utils";
 import {
   AutomatedVotingOptionCurrentDelegation,
   AutomatedVotingOptionDelegationId,
@@ -40,8 +40,8 @@ export const AutomatedVotingOptions = ({
   txHash,
 }: AutomatedVotingOptionsProps) => {
   const { t } = useTranslation();
-  // TODO: Get network metrics from useAppContext
-  const { networkMetrics, fetchNetworkMetrics } = useGetNetworkMetrics();
+  const { networkTotalStake, fetchNetworkTotalStake } =
+    useGetNetworkTotalStake();
 
   const [isOpen, setIsOpen] = useState<boolean>(false);
 
@@ -57,7 +57,7 @@ export const AutomatedVotingOptions = ({
     delegationInProgress === AutomatedVotingOptionDelegationId.no_confidence;
 
   useEffect(() => {
-    fetchNetworkMetrics();
+    fetchNetworkTotalStake();
   }, []);
 
   useEffect(() => {
@@ -128,8 +128,10 @@ export const AutomatedVotingOptions = ({
                 : t("dRepDirectory.abstainCardDefaultTitle")
             }
             votingPower={
-              networkMetrics
-                ? correctVoteAdaFormat(networkMetrics?.alwaysAbstainVotingPower)
+              networkTotalStake
+                ? correctDRepDirectoryFormat(
+                    networkTotalStake?.alwaysAbstainVotingPower,
+                  )
                 : ""
             }
             transactionId={
@@ -169,9 +171,9 @@ export const AutomatedVotingOptions = ({
                 : t("dRepDirectory.noConfidenceDefaultTitle")
             }
             votingPower={
-              networkMetrics
-                ? correctVoteAdaFormat(
-                    networkMetrics?.alwaysNoConfidenceVotingPower,
+              networkTotalStake
+                ? correctDRepDirectoryFormat(
+                    networkTotalStake?.alwaysNoConfidenceVotingPower,
                   )
                 : ""
             }