+++ govtool/frontend/src/models/api.ts
@@ -78,25 +78,39 @@ export type TransactionStatus = {
     | [];
 };
 
-export type NetworkMetrics = {
+export enum Network {
+  samchonet = "sanchonet",
+  preview = "preview",
+  testnet = "testnet",
+  preprod = "preprod",
+  mainnet = "mainnet",
+}
+
+export type NetworkInfo = {
   currentTime: string;
   currentEpoch: number;
   currentBlock: number;
+  networkName: Network;
+};
+
+export type NetworkTotalStake = {
+  totalStakeControlledByDReps: number;
+  totalStakeControlledBySPOs: number;
+  alwaysAbstainVotingPower: number;
+  alwaysNoConfidenceVotingPower: number;
+};
+
+export type NetworkMetrics = {
   uniqueDelegators: number;
   totalDelegations: number;
   totalGovernanceActions: number;
   totalDRepVotes: number;
   totalRegisteredDReps: number;
   totalDRepDistr: number;
-  totalStakeControlledByDReps: number;
-  totalStakeControlledBySPOs: number;
   totalActiveDReps: number;
   totalInactiveDReps: number;
   totalActiveCIP119CompliantDReps: number;
   totalRegisteredDirectVoters: number;
-  alwaysAbstainVotingPower: number;
-  alwaysNoConfidenceVotingPower: number;
-  networkName: string;
   noOfCommitteeMembers: number;
   quorumNumerator: number;
   quorumDenominator: number;