+++ tests/govtool-frontend/playwright/lib/types.ts
@@ -9,6 +9,13 @@ export type KuberValue = {
   [policyId: string]: Record<string, BigInt | number> | BigInt | number;
 };
 
+export interface PaginatedLiveProposal {
+  page: number;
+  pageSize: number;
+  total: number;
+  elements: IProposal[];
+}
+
 export interface IProposal {
   id: string;
   txHash: string;
@@ -90,6 +97,16 @@ export enum GovernanceActionType {
   NoConfidence = "NoConfidence",
   NewCommittee = "NewCommittee",
   UpdatetotheConstitution = "NewConstitution",
+}
+
+export enum outcomeType {
+  NewConstitution = "New Constitution",
+  NewCommittee = "Update Committee",
+  HardForkInitiation = "Hard-Fork Initiation",
+  NoConfidence = "Motion of no Confidence",
+  InfoAction = "Info Action",
+  TreasuryWithdrawals = "Treasury Withdrawals",
+  ParameterChange = "Protocol Parameter Change",
 }
 
 export enum FullGovernanceDRepVoteActionsType {
@@ -250,4 +267,52 @@ export interface imageObject {
   "@type": "ImageObject";
   contentUrl: string;
   sha256: string;
+}
+
+export interface outcomeProposal {
+  id: string;
+  tx_hash: string;
+  index: string;
+  type: string;
+  yes_votes: string;
+  no_votes: string;
+  abstain_votes: string;
+  description: any;
+  expiry_date: string;
+  expiration: number;
+  time: string;
+  epoch_no: number;
+  url: string;
+  data_hash: string;
+  title: string | null;
+  abstract: string | null;
+  motivation?: string | null;
+  rationale?: string | null;
+  pool_yes_votes?: string;
+  pool_no_votes?: string;
+  pool_abstain_votes?: string;
+  cc_yes_votes?: string;
+  cc_no_votes?: string;
+  cc_abstain_votes?: string;
+  proposal_params: EpochParams | null;
+}
+
+export interface outcomeMetadata {
+  metadataStatus: string;
+  metadataValid: boolean;
+  data: outcomeMetadataBody;
+}
+
+interface outcomeMetadataBody {
+  abstract: string;
+  motivation: "string";
+  rationale: string;
+  title: string;
+}
+
+export interface InvalidMetadataType {
+  type: string;
+  reason: string;
+  url: string;
+  hash: string;
 }